uucore: Move fast_inc functions from seq
A new fast-inc feature, to be used by seq and cat.
